IA-3819: regroup network code by domain
- remove domain specific functions from utils/requests.js
- convert remaining code to ts
- re-organize folders to have network code in or close to components
- convert some js components to TS (loose typing)
