+++ hat/assets/js/apps/Iaso/domains/instances/components/InstancesFilesListComponent.tsx
@@ -11,9 +11,11 @@ import VideosList from '../../../components/files/VideosListComponent';
 import InstancePopover from './InstancePopoverComponent';
 
 import { SortedFiles, sortFilesType } from '../../../utils/filesUtils';
-import { fetchInstanceDetail } from '../../../utils/requests';
 import MESSAGES from '../messages';
 import { Instance, ShortFile } from '../types/instance';
+import { getRequest } from '../../../libs/Api';
+import { openSnackBar } from '../../../components/snackBars/EventDispatcher';
+import { errorSnackBar } from '../../../constants/snackBars';
 
 const minTabHeight = 'calc(100vh - 500px)';
 
@@ -39,6 +41,15 @@ const useStyles = makeStyles(theme => ({
         backgroundColor: 'white',
     },
 }));
+
+const fetchInstanceDetail = instanceId =>
+    getRequest(`/api/instances/${instanceId}/`)
+        .then(instance => instance)
+        .catch(error => {
+            openSnackBar(errorSnackBar('fetchInstanceError', null, error));
+            console.error('Error while fetching instance detail:', error);
+        });
+
 type Props = {
     instanceDetail?: Instance;
     files: ShortFile[];
