+++ hat/assets/js/apps/Iaso/utils/requests.ts
@@ -1,0 +1,21 @@
+export const cleanupParams = params => {
+    const copy = { ...params };
+    Object.keys(params).forEach(key => {
+        if (copy[key] === undefined) {
+            delete copy[key];
+        }
+    });
+    return copy;
+};
+
+export const formatParams = params => {
+    const copy = cleanupParams(params);
+    if (params.pageSize) {
+        copy.limit = params.pageSize;
+        delete copy.pageSize;
+    }
+    if (params.accountId) {
+        delete copy.accountId;
+    }
+    return copy;
+};
